Removed extra functions for each proxy site which use identical divs
Removed extra functions for each proxy site to one, as these sites that are scraped have the same layout and divs for table containing proxies. Put URL as parameter passed to makesoup function in the scrapeproxies function in the menu. Also added an error if user selects incorrect number in menu.